Javadoc corrections; remove extraneous semi-colons
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@755098 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/vfs/provider/tar/TarInputStream.java
@@ -307,7 +307,7 @@ class TarInputStream
         }
         else
         {
-            return (int)m_oneBuf[ 0 ];
+            return m_oneBuf[ 0 ];
         }
     }
